[FE] REFACTOR: 테이블 컴포넌트 하나로 통일 #956
