+++ frontend/src/App.tsx
@@ -6,12 +6,11 @@ import HomePage from "@/pages/HomePage";
 import MainPage from "@/pages/MainPage";
 import LoadingAnimation from "./components/Common/LoadingAnimation";
 import LogPage from "@/pages/LogPage";
-import AdminInfo from "./components/AdminChart/AdminInfo";
+import AdminInfo from "./components/AdminInfo/AdminInfo";
 const NotFoundPage = lazy(() => import("@/pages/NotFoundPage"));
 const LoginFailurePage = lazy(() => import("@/pages/LoginFailurePage"));
 const AdminLayout = lazy(() => import("@/pages/admin/AdminLayout"));
 const AdminLoginPage = lazy(() => import("@/pages/admin/AdminLoginPage"));
-const AdminHomePage = lazy(() => import("@/pages/admin/AdminHomePage"));
 const AdminLogPage = lazy(() => import("@/pages/admin/AdminLogPage"));
 const SearchPage = lazy(() => import("@/pages/admin/SearchPage"));
 const AdminLoginFailurePage = lazy(
